Detect NetBSD and add casts to quell complaints with toupper() and isspace().

+++ src/curses/ux_init.c
@@ -835,7 +835,7 @@ static int getconfig(char *configfile)
 			;
 
 		/* Remove trailing whitespace and newline */
-		for (num = strlen(varname) - 1; (ssize_t) num >= 0 && isspace(varname[num]); num--)
+		for (num = strlen(varname) - 1; (ssize_t) num >= 0 && isspace((int) varname[num]); num--)
 		;
 		varname[num+1] = 0;
 
@@ -846,9 +846,9 @@ static int getconfig(char *configfile)
 		}
 
 		/* Find end of variable name */
-		for (num = 0; varname[num] != 0 && !isspace(varname[num]) && num < LINELEN; num++);
+		for (num = 0; varname[num] != 0 && !isspace((int) varname[num]) && num < LINELEN; num++);
 
-		for (num2 = num; isspace(varname[num2]) && num2 < LINELEN; num2++);
+		for (num2 = num; isspace((int) varname[num2]) && num2 < LINELEN; num2++);
 
 		/* Find the beginning of the value */
 		strncpy(value, &varname[num2], LINELEN);
@@ -962,7 +962,7 @@ static int getbool(char *value)
 
 	/* Be case-insensitive */
 	for (num = 0; value[num] !=0; num++)
-		value[num] = tolower(value[num]);
+		value[num] = tolower((int) value[num]);
 
 	if (strncmp(value, "y", 1) == 0)
 		return TRUE;
@@ -990,7 +990,7 @@ static int getcolor(char *value)
 
 	/* Be case-insensitive */
 	for (num = 0; value[num] !=0; num++)
-		value[num] = tolower(value[num]);
+		value[num] = tolower((int) value[num]);
 
 	if (strcmp(value, "black") == 0)
 		return BLACK_COLOUR;
@@ -1038,7 +1038,7 @@ static int geterrmode(char *value)
 
         /* Be case-insensitive */
 	for (num = 0; value[num] !=0; num++)
-		value[num] = tolower(value[num]);
+		value[num] = tolower((int) value[num]);
 
 	if (strcmp(value, "never") == 0)
 		return ERR_REPORT_NEVER;
